fix: add ibis-pandas output to feedstock
the pandas backend is currently always available because it's a core
dependency, but for the sake of consistency, adding it here so we don't
have to explain on the docs why there is no `ibis-pandas` conda-forge
package.

The output here is an exact copy of the `ibis-duckdb` output, since we
still want DuckDB as it is the default backend.